requirements.txt
Flask
pandas
numpy
pytorch